refactor(161): move puzzle validation into validation package
