+++ app/src/main/java/org/cescfe/numpairs/domain/puzzle/model/Puzzle.kt
@@ -1,9 +1,9 @@
 package org.cescfe.numpairs.domain.puzzle.model
 
-import org.cescfe.numpairs.domain.puzzle.hasIncorrectTiles
-import org.cescfe.numpairs.domain.puzzle.hasInvalidStripEntryUsage
-import org.cescfe.numpairs.domain.puzzle.hasMismatchedSumProductPairings
-import org.cescfe.numpairs.domain.puzzle.hasMissingResolvedTileAssignments
+import org.cescfe.numpairs.domain.puzzle.validation.hasIncorrectTiles
+import org.cescfe.numpairs.domain.puzzle.validation.hasInvalidStripEntryUsage
+import org.cescfe.numpairs.domain.puzzle.validation.hasMismatchedSumProductPairings
+import org.cescfe.numpairs.domain.puzzle.validation.hasMissingResolvedTileAssignments
 
 data class Puzzle(val board: Board, val strip: Strip) {
     init {
