+++ app/src/main/java/org/cescfe/numpairs/domain/puzzle/validation/LivePuzzleRuleConflict.kt
@@ -1,4 +1,4 @@
-package org.cescfe.numpairs.domain.puzzle
+package org.cescfe.numpairs.domain.puzzle.validation
 
 import org.cescfe.numpairs.domain.puzzle.model.Board
 import org.cescfe.numpairs.domain.puzzle.model.Expression
