+++ app/src/test/java/org/cescfe/numpairs/domain/puzzle/LivePuzzleRuleConflictTest.kt
@@ -11,6 +11,11 @@ import org.cescfe.numpairs.domain.puzzle.model.Strip
 import org.cescfe.numpairs.domain.puzzle.model.StripItem
 import org.cescfe.numpairs.domain.puzzle.support.hiddenTile
 import org.cescfe.numpairs.domain.puzzle.support.withTile
+import org.cescfe.numpairs.domain.puzzle.validation.LivePuzzleRuleConflict
+import org.cescfe.numpairs.domain.puzzle.validation.LiveStripEntryOperatorUsage
+import org.cescfe.numpairs.domain.puzzle.validation.liveRuleConflictsByStripEntryOperator
+import org.cescfe.numpairs.domain.puzzle.validation.liveRuleConflictsByTile
+import org.cescfe.numpairs.domain.puzzle.validation.liveRuleConflictsForCandidate
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertTrue
 import org.junit.Test
