+++ app/src/test/java/org/cescfe/numpairs/domain/puzzle/StripEntryUsageTest.kt
@@ -4,6 +4,8 @@ import org.cescfe.numpairs.domain.puzzle.model.Operator
 import org.cescfe.numpairs.domain.puzzle.model.TileResolutionState
 import org.cescfe.numpairs.domain.puzzle.support.TileAssignment
 import org.cescfe.numpairs.domain.puzzle.support.knownPuzzleWithAssignments
+import org.cescfe.numpairs.domain.puzzle.validation.hasInvalidStripEntryUsage
+import org.cescfe.numpairs.domain.puzzle.validation.hasMismatchedSumProductPairings
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertTrue
 import org.junit.Test
